+++ grafana_client/elements/__init__.py
@@ -2,6 +2,7 @@ from .admin import Admin
 from .annotations import Annotations
 from .base import Base
 from .dashboard import Dashboard
+from .dashboard_versions import DashboardVersions
 from .datasource import Datasource
 from .folder import Folder
 from .health import Health